Revert to the previous state

Restored to the latest state due to data loss.


Organize directory structure due to inconsistencies


Made changes for the pull request


change can_all_node to cabot_can_node


Remove unnecessary variables
