+++ cabot_can/CMakeLists.txt
@@ -10,11 +10,34 @@ find_package(rclcpp REQUIRED)
 find_package(sensor_msgs REQUIRED)
 find_package(std_msgs REQUIRED)
 find_package(std_srvs REQUIRED)
+find_package(diagnostic_updater REQUIRED)
+add_executable(cabot_can_node
+  src/cabot_can_node.cpp
+)
+ament_target_dependencies(cabot_can_node
+  diagnostic_updater
+  rclcpp
+  std_msgs
+  sensor_msgs
+  std_srvs
+)
 
-add_executable(can_all_node src/can_all_node.cpp)
-ament_target_dependencies(can_all_node rclcpp std_msgs sensor_msgs std_srvs)
+if(BUILD_TESTING)
+  find_package(ament_cmake_copyright REQUIRED)
+  find_package(ament_cmake_cpplint REQUIRED)
+  find_package(ament_cmake_lint_cmake REQUIRED)
+  find_package(ament_cmake_uncrustify REQUIRED)
+  find_package(ament_cmake_xmllint REQUIRED)
 
-install(TARGETS can_all_node
+  set(MAX_LINE_LENGTH 200)
+  ament_copyright()
+  ament_cpplint(MAX_LINE_LENGTH ${MAX_LINE_LENGTH})
+  ament_lint_cmake()
+  ament_uncrustify(MAX_LINE_LENGTH ${MAX_LINE_LENGTH})
+  ament_xmllint()
+endif()
+
+install(TARGETS cabot_can_node
   DESTINATION lib/${PROJECT_NAME})
 
 ament_package()
