+++ power_controller/power_controller_msgs/msg/BatteryArray.msg
